+++ R/Reintegrate.R
@@ -33,7 +33,7 @@
 #'   "extdata", "example1sec.csv", package = "AGread"
 #' )
 #'
-#' ag <- read_AG_counts(test_file, skip = 11)
+#' ag <- read_AG_counts(test_file, header = TRUE)
 #'
 #' # Old Method:
 #'